+++ src/main/resources/help/content/releasenotes.md
@@ -1,6 +1,6 @@
 #### Version 9.0.2
 ##### Resolved issues
-* Resolved an issue that could caused ${solution_name} to fail when using existing image inspector services when given a target docker .tar file that resided outside the directory shared with the image inspector container(s).
+* Resolved an issue that could cause ${solution_name} to fail when using existing image inspector services when given a target docker .tar file that resided outside the directory shared with the image inspector container(s).
 
 #### Version 9.0.1
 ##### Resolved issues